refactor: user handler init
